requirements.txt
aws-cdk.cdk
aws-cdk.aws-events
aws-cdk.aws-events-targets
aws-cdk.aws_lambda
boto3
json-logging
pytz
pytest